Move get_expired_tokens to push_notifications.utils
